refactor: OutfitPostGenderSelect 재사용 컴포넌트로 분리
