+++ src/features/outfits/components/OutfitPostGenderSelect/index.ts
@@ -1,0 +1,3 @@
+import OutfitPostGenderSelect from './OutfitPostGenderSelect';
+
+export default OutfitPostGenderSelect;
